+++ src/CScsi_Background_Operation_Log.cpp
@@ -183,10 +183,10 @@ void CScsiOperationLog::process_Background_Operations_Data(JSONNODE *operationDa
     temp << "0x" << std::hex << std::setfill('0') << std::setw(4) << m_Operation->paramCode;
 	json_push_back(operationInfo, json_new_a("Background Operation Parameter Code", temp.str().c_str()));
     temp.clear();
-    temp << "0x" << std::hex << std::setfill('0') << std::setw(2) << m_Operation->paramControlByte;
+    temp << "0x" << std::hex << std::setfill('0') << std::setw(2) << static_cast<uint16_t>(m_Operation->paramControlByte);
 	json_push_back(operationInfo, json_new_a("Background Operation Control Byte ", temp.str().c_str()));
     temp.clear();
-    temp << "0x" << std::hex << std::setfill('0') << std::setw(2) << m_Operation->paramLength;
+    temp << "0x" << std::hex << std::setfill('0') << std::setw(2) << static_cast<uint16_t>(m_Operation->paramLength);
 	json_push_back(operationInfo, json_new_a("Background Operation Length ", temp.str().c_str()));
 
 	get_Background_Operations_status(&myStr);
